devman_lessons/elite_vine_sell/requirements.txt
et-xmlfile==1.1.*
Jinja2==3.1.*
numpy==1.23.*
openpyxl==3.0.*
pandas==1.4.*
python-dotenv==0.20.*